pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loanlens"
version = "0.1.0"
description = "Personal home loan management CLI utility"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "click>=8",
  "numpy-financial>=1.0",
  "pandas>=2",
  "prompt_toolkit>=3",
  "pydantic>=2",
  "python-dateutil>=2.9",
  "rich>=13",
  "typer>=0.12",
]

[project.optional-dependencies]
pdf = ["reportlab>=4"]
dev = [
  "mypy>=1.10",
  "pre-commit>=3.7",
  "pytest>=8",
  "pytest-cov>=5",
  "pytest-mock>=3.14",
  "ruff>=0.5",
]

[project.scripts]
loanlens = "loanlens.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["loanlens*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ANN", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "ANN001", "ANN202"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]

